requirements.txt
beautifulsoup4==4.13.3
certifi==2025.1.31
cffi==1.17.1
charset-normalizer==3.4.1
cryptography==44.0.2
frozendict==2.4.6
idna==3.10
multitasking==0.0.11
mysql-connector-python==9.2.0
numpy==2.2.4
pandas==2.2.3
peewee==3.17.9
platformdirs==4.3.7
pycparser==2.22
pyotp==2.9.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2025.1
requests==2.32.3
robin-stocks==3.3.1
six==1.17.0
soupsieve==2.6
typing_extensions==4.13.0
tzdata==2025.1
urllib3==2.3.0
yfinance==0.2.55
